requirements.txt
addict==2.4.0
albumentations==2.0.8
apex==0.9.10dev
cityscapesscripts==2.2.4
einops==0.8.1
evo==1.31.1
fire==0.7.1
flow_vis==0.1
gpytorch==1.13
gradio==5.45.0
gsplat==1.5.3
h5py==3.14.0
imageio==2.37.0
ipython==8.12.3
joblib==1.4.2
lietorch==0.3
lmdb==1.7.3
loguru==0.7.3
lxml==5.4.0
lz4==4.4.4
matplotlib==3.10.6
mc==0.0
mock==5.2.0
model_archiver==1.0.3
munch==4.0.0
nerfview==0.1.3
numpy==1.24.3
omegaconf==2.3.0
open3d==0.15.1
oss2==2.19.1
packaging==25.0
pandas==2.3.2
Pillow==11.3.0
plyfile==1.1.2
popart==0.1.1
prettytable==3.16.0
psutil==6.1.0
psutil==5.9.0
pytest==8.4.2
pytorch3d==0.7.7
pytorch_sphinx_theme==0.0.19
PyYAML==6.0.2
PyYAML==6.0.2
regex==2025.9.1
renderpy==1.0.0
Requests==2.32.5
rerun==1.0.30
rich==14.1.0
scikit_learn==1.7.2
scipy==1.16.1
seaborn==0.13.2
setuptools==73.0.1
skimage==0.0
spaces==0.40.3
Sphinx==8.2.3
tensorboardX==2.6.4
tifffile==2024.6.18
timm==1.0.19
torch==2.0.1
torch_kmeans==0.2.0
torch_tb_profiler==0.4.3
torchmetrics==1.7.2
torchvision==0.15.2
tqdm==4.67.1
trimesh==4.6.10
ts==0.5.1
turbojpeg==0.0.2
viser==1.0.8
xformers==0.0.32.post2
yapf==0.43.0
